requirements.txt
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
clip==1.0
Flask==2.2.2
ftfy==6.1.1
idna==3.4
importlib-metadata==5.1.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
numpy==1.23.5
Pillow==9.3.0
regex==2022.10.31
requests==2.28.1
torch==1.13.0
torchvision==0.14.0
tqdm==4.64.1
typing_extensions==4.4.0
urllib3==1.26.13
wcwidth==0.2.5
Werkzeug==2.2.2
zipp==3.11.0
certifi==2022.9.24
charset-normalizer==2.1.1
docopt==0.6.2
idna==3.4
requests==2.28.1
urllib3==1.26.13
yarg==0.1.9
